+++ web/src/providers/InnerAuthProvider.js
@@ -1,0 +1,7 @@
+import React from 'react'
+import { useOuterAuthContext } from './OuterAuthProvider'
+
+export const InnerAuthProvider = ({ children }) => {
+    const { getCurrentUser } = useOuterAuthContext()
+    return getCurrentUser() ? children : 'loading ...'
+}