+++ web/src/services/leaderboard-service.js
@@ -1,8 +1,8 @@
-import { useAuthContext } from "@providers/AuthProvider"
+import { useOuterAuthContext } from "@providers/OuterAuthProvider"
 import axios from "axios"
 
 export const useLeaderboardService = () => {
-    const authContext = useAuthContext()
+    const authContext = useOuterAuthContext()
 
     const fetchUsersLeaderboard = async () => {
         const response = await axios.get(`/leaderboard/users`, {